IPRT/isovfs.cpp: Use rock ridge information for locating and listing files on ISO-9660 volumes. [build fix] bugref:9781
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@94282 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/fs/isovfs.cpp
@@ -2903,7 +2903,7 @@ static void rtFsIsoDirShrd_ParseRockRidgeData(PRTFSISOVOL pVol, PRTFSISOROCKINFO
                     }
                     Assert(offDst < sizeof(pParseInfo->szLinkTarget));
                     pParseInfo->szLinkTarget[offDst] = '\0';
-                    pParseInfo->cchLinkTarget        = offDst;
+                    pParseInfo->cchLinkTarget        = (uint16_t)offDst;
                 }
                 break;
